+++ package.json
@@ -24,6 +24,7 @@
     "eslint-plugin-react": "7.4.0",
     "extract-text-webpack-plugin": "3.0.2",
     "file-loader": "1.1.5",
+    "firebase": "^5.0.4",
     "fs-extra": "3.0.1",
     "html-webpack-plugin": "2.29.0",
     "jest": "20.0.4",
@@ -36,6 +37,7 @@
     "react": "^16.4.0",
     "react-dev-utils": "^5.0.1",
     "react-dom": "^16.4.0",
+    "react-firebase": "^2.2.8",
     "recharts": "^1.0.0-beta.10",
     "resolve": "1.6.0",
     "sass-loader": "^7.0.2",
