+++ public/index.html
@@ -19,20 +19,5 @@
     <div id="questions" class="questions">
     </div>
     
-    
-    <script src="https://www.gstatic.com/firebasejs/5.0.4/firebase.js"></script>
-    <script>
-      // Initialize Firebase
-      var config = {
-        apiKey: "AIzaSyBO7y0d96shvYCtIt3blyVxo7JO03likZ4",
-        authDomain: "wkmb-42701.firebaseapp.com",
-        databaseURL: "https://wkmb-42701.firebaseio.com",
-        projectId: "wkmb-42701",
-        storageBucket: "",
-        messagingSenderId: "327208320794"
-      };
-      firebase.initializeApp(config);
-    </script>
-    <script type="text/javascript" src="firescripts.js"></script>
   </body>
 </html>
